backend/requirements.txt
# TECHGURU ElevateCRM Backend Dependencies

# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.9

# Data Validation and Serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Authentication and Security
PyJWT[crypto]==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client and External APIs
httpx==0.25.2
requests==2.31.0

# Background Jobs and Caching
celery[redis]==5.3.4
redis==4.6.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
